# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teste prático - PHP",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/usuarios": {
            "get": {
                "tags": [
                    "usuario"
                ],
                "summary": "listar_usuarios",
                "description": "Retornar todos os cadastros",
                "operationId": "listar_usuarios",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tudo certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . ",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Erro interno!",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/usuario": {
            "post": {
                "tags": [
                    "usuario"
                ],
                "summary": "Cadastrar Novo Usuario",
                "operationId": "new_usuario",
                "requestBody": {
                    "description": "Campos para cadastros",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email",
                                    "password",
                                    "nivel"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "nome",
                                        "format": "text",
                                        "example": "Marcos"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "marcoslopesg7@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tudo Certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . "
                    },
                    "422": {
                        "description": "Entradas inválidas, campos obrigatórios! "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/usuario/{id}": {
            "put": {
                "tags": [
                    "usuario"
                ],
                "summary": "Alterar cadastro do usuario",
                "operationId": "editar_usuario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Campos para cadastros",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email",
                                    "password",
                                    "nivel"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "nome",
                                        "format": "text",
                                        "example": "Marcos"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "marcoslopesg7@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tudo Certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . "
                    },
                    "422": {
                        "description": "Entradas inválidas, campos obrigatórios! "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "usuario"
                ],
                "summary": "Excluir cadastro de um usuario",
                "operationId": "excluir_usuario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tudo Certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "Efetuar login ",
                "operationId": "login_usuario",
                "requestBody": {
                    "description": "Campos para cadastros",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "marcoslopesg7@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tudo Certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . "
                    },
                    "422": {
                        "description": "Entradas inválidas, campos obrigatórios! "
                    }
                }
            }
        },
        "/api/v1/login-check": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "Verificar Token",
                "description": "Verificar Token",
                "operationId": "login_usuario_check",
                "responses": {
                    "200": {
                        "description": "Tudo certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . ",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Erro interno!",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearerAuth"
            }
        }
    ]
}